requirements.txt
requests
python3-bs4
termcolor